Move url serialiser to core module
